+++ src/plugins/header/options/options.ts
@@ -1,4 +1,4 @@
-import {bind} from '@exadel/esl/modules/esl-utils/decorators/bind';
+import {listen} from '@exadel/esl/modules/esl-utils/decorators/listen';
 
 import {UIPPlugin} from '../../../core/base/plugin';
 import {OptionConfig, UIPOption} from './option/option';
@@ -35,24 +35,12 @@ export class UIPOptions extends UIPPlugin {
 
   protected connectedCallback() {
     super.connectedCallback();
-    this.bindEvents();
     this.render();
   }
 
   protected disconnectedCallback() {
     this.innerHTML = '';
-    this.unbindEvents();
     super.disconnectedCallback();
-  }
-
-  protected bindEvents() {
-    this.addEventListener('esl:uip:option:changed', this._onOptionClick);
-    this.root?.addEventListener('uip:configchange', this._onRootConfigChange);
-  }
-
-  protected unbindEvents() {
-    this.removeEventListener('esl:uip:option:changed', this._onOptionClick);
-    this.root?.removeEventListener('uip:configchange', this._onRootConfigChange);
   }
 
   protected render() {
@@ -62,13 +50,13 @@ export class UIPOptions extends UIPPlugin {
     this.options.forEach(option => this.append(option));
   }
 
-  @bind
+  @listen({event: 'uip:configchange', target: '::parent(.uip-root)'})
   protected _onRootConfigChange(e: CustomEvent) {
     const option = this.options.get(e.detail.attribute);
     option?.toggleState(e.detail.value !== null);
   }
 
-  @bind
+  @listen('click')
   protected _onOptionClick(e: Event) {
     e.stopPropagation();
     const option = e.target as UIPOption;